# receiver/prometheus: Populate scope attributes from labels instead of otel_scope_info