--- a/receiver/prometheusreceiver/internal/metricfamily.go
+++ b/receiver/prometheusreceiver/internal/metricfamily.go
@@ -361,9 +361,7 @@ func populateAttributes(mType pmetric.MetricType, ls labels.Labels, dest pcommon
 			return
 		}
 
-		// When removeScopeInfo feature gate is enabled, filter out otel_scope_ prefixed labels
-		// as they are now extracted as scope attributes instead of datapoint attributes
-		if RemoveScopeInfoGate.IsEnabled() && strings.HasPrefix(l.Name, "otel_scope_") {
+		if strings.HasPrefix(l.Name, "otel_scope_") {
 			return
 		}