… .chloggen/prometheusreceiver_scope_attributes_spec_compliance.yaml
… receiver/prometheusreceiver/README.md
… receiver/prometheusreceiver/internal/metricfamily_test.go
… receiver/prometheusreceiver/metrics_receiver_labels_test.go