This PR is a work in progress. It cannot handle multiple metrics that should be treated as having the same Scope.

#### Description
This PR is a PoC for the spec change described at https://github.com/open-telemetry/opentelemetry-specification/issues/4223.

In the specification change, it's proposed that Prometheus exporters stop generating the metric `otel_scope_info` with all the scope attributes and instead add all scope information as labels, turning the Scope information "identifying".

For the receiver, this means that Scope should no longer be populated from the `otel_scope_info` but should instead be looked for metrics with the same labels prefixed with `otel_scope_`.

For users who still haven't updated to newer versions of the SDK that adhere to the spec change, ignoring the `otel_scope_info` metric would be a breaking change from this receiver's perspective. For that reason, this change is being made through feature flags.

If an exporter exposes `otel_scope_info` and the feature gate `receiver.prometheusreceiver.RemoteScopeInfo` is enabled, then `otel_scope_info` will be transformed into a metric like all others instead of being cached and used to generate ScopeMetrics information