@ArthurSens 2025-05-26
Manual tests are showing that I broke the current behavior that extracts attributes from `otel_scope_info` metrics and when when the feature gate is enabled I'm getting nil pointers xD

Back to draft