@ArthurSens 2025-08-12
Ok, I finally found some time to get back to this. I ended up fixing the previous problems by introducing an interface. I was afraid this would introduce some performance penalty, so I implemented a benchmark before proceeding. CPU profiles show increase in runtime operations from ~20s to ~22 between benchmarks.

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/receiver/prometheusreceiver/internal
cpu: Apple M2 Pro
                                       │ internal/bench-main.txt │         benchmark-new.txt          │
                                       │         sec/op          │   sec/op     vs base               │
TransactionAppendAndCommit/Floats-2                  2.668µ ± 2%   2.872µ ± 2%   +7.63% (p=0.002 n=6)
TransactionAppendAndCommit/Histogram-2               6.550µ ± 1%   7.263µ ± 2%  +10.88% (p=0.002 n=6)
geomean                                              4.181µ        4.567µ        +9.24%

                                       │ internal/bench-main.txt │         benchmark-new.txt          │
                                       │          B/op           │     B/op      vs base              │
TransactionAppendAndCommit/Floats-2                 2.410Ki ± 0%   2.555Ki ± 0%  +6.00% (p=0.002 n=6)
TransactionAppendAndCommit/Histogram-2              5.465Ki ± 0%   5.918Ki ± 0%  +8.29% (p=0.002 n=6)
geomean                                             3.629Ki        3.888Ki       +7.14%

                                       │ internal/bench-main.txt │         benchmark-new.txt         │
                                       │        allocs/op        │ allocs/op   vs base               │
TransactionAppendAndCommit/Floats-2                   22.00 ± 0%   26.00 ± 0%  +18.18% (p=0.002 n=6)
TransactionAppendAndCommit/Histogram-2                48.00 ± 0%   64.00 ± 0%  +33.33% (p=0.002 n=6)
geomean                                               32.50        40.79       +25.53%

```

<img width="1727" height="433" alt="image" src="https://github.com/user-attachments/assets/b2cf4dae-c3b6-474d-a5ec-2b228482509b" />


Not sure if we're in the position to care too much about the performance, but please let me know if I could be taking another approach!