@ArthurSens 2025-08-13
> I think we should proceed with whatever you have working, and open an issue with your profiles, etc to look into optimizing that code path.

Once the `otel_scope_info` is fully deprecated, we could remove the interface again and then the compiler will be able to inline a lot of the codepath again 🤔