@ArthurSens 2025-09-01
> I imagine people will have a mix of clients for some time. Some will use the new scope attributes on metrics, and others will still be using otel_scope_info. I think we should **always** add otel_scope_foo attributes as scope attributes, and use the feature gate to remove special handling of otel_scope_info.

Hmmm, good point. Let me change that